+++ Libs/StateMachine.h
@@ -12,7 +12,7 @@ typedef struct T_StateMachine StateMachine;
 typedef struct
 {
     unsigned int m_Prio;
-    void (*m_Callback)(u_int32_t _MSTime, void* _Context);
+    void (*m_Callback)(UInt64 _MSTime, void* _Context);
     void* m_Context;
 
 } StateMachine_Task;
@@ -29,7 +29,7 @@ struct T_StateMachine
 int StateMachine_InitializePtr(StateMachine** _StateMachinePtr);
 int StateMachine_Initialize(StateMachine* _StateMachine);
 
-int StateMachine_CreateTask(StateMachine* _StateMachine, unsigned int _Prio, const char* _Name, void (*_Callback)(u_int32_t _MSTime, void* _Context), void* _Context, StateMachine_Task** _TaskPtr);
+int StateMachine_CreateTask(StateMachine* _StateMachine, unsigned int _Prio, const char* _Name, void (*_Callback)(UInt64 _MSTime, void* _Context), void* _Context, StateMachine_Task** _TaskPtr);
 
 void StateMachine_Dispose(StateMachine* _StateMachine);
